+++ crates/sncast/tests/e2e/account/deploy.rs
@@ -23,7 +23,7 @@ use test_case::test_case;
 #[test_case(&ARGENT_CLASS_HASH.into_hex_string(), "argent"; "argent_class_hash")]
 #[test_case(&BRAAVOS_CLASS_HASH.into_hex_string(), "braavos"; "braavos_class_hash")]
 #[tokio::test]
-pub async fn test_happy_case(class_hash: &str, account_type: &str) {
+pub async fn test_happy_case_eth(class_hash: &str, account_type: &str) {
     let tempdir = create_account(false, class_hash, account_type).await;
     let accounts_file = "accounts.json";
 
@@ -39,6 +39,8 @@ pub async fn test_happy_case(class_hash: &str, account_type: &str) {
         "my_account",
         "--max-fee",
         "99999999999999999",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -54,8 +56,167 @@ pub async fn test_happy_case(class_hash: &str, account_type: &str) {
     assert!(stdout_str.contains("transaction_hash"));
 
     let contents = fs::read_to_string(tempdir.path().join(accounts_file)).unwrap();
-    let items: serde_json::Value =
-        serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
+    let items: Value = serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
+    assert_eq!(items["alpha-sepolia"]["my_account"]["deployed"], true);
+}
+
+#[tokio::test]
+pub async fn test_happy_case_v1() {
+    let tempdir = create_account(false, &OZ_CLASS_HASH.into_hex_string(), "oz").await;
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--json",
+        "account",
+        "deploy",
+        "--name",
+        "my_account",
+        "--max-fee",
+        "99999999999999999",
+        "--version",
+        "v1",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+    let bdg = snapbox.assert();
+
+    let hash = get_transaction_hash(&bdg.get_output().stdout);
+    let receipt = get_transaction_receipt(hash).await;
+
+    assert!(matches!(receipt, DeployAccount(_)));
+
+    let stdout_str = bdg.as_stdout();
+    assert!(stdout_str.contains("account deploy"));
+    assert!(stdout_str.contains("transaction_hash"));
+
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file)).unwrap();
+    let items: Value = serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
+    assert_eq!(items["alpha-sepolia"]["my_account"]["deployed"], true);
+}
+
+#[test_case(DEVNET_OZ_CLASS_HASH_CAIRO_0, "oz"; "cairo_0_class_hash")]
+#[test_case(&OZ_CLASS_HASH.into_hex_string(), "oz"; "cairo_1_class_hash")]
+#[test_case(&ARGENT_CLASS_HASH.into_hex_string(), "argent"; "argent_class_hash")]
+#[test_case(&BRAAVOS_CLASS_HASH.into_hex_string(), "braavos"; "braavos_class_hash")]
+#[tokio::test]
+pub async fn test_happy_case_strk(class_hash: &str, account_type: &str) {
+    let tempdir = create_account(false, class_hash, account_type).await;
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--json",
+        "account",
+        "deploy",
+        "--name",
+        "my_account",
+        "--fee-token",
+        "strk",
+        "--max-gas",
+        "1000",
+        "--max-gas-unit-price",
+        "100000000000",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+    let bdg = snapbox.assert();
+
+    let hash = get_transaction_hash(&bdg.get_output().stdout);
+    let receipt = get_transaction_receipt(hash).await;
+
+    assert!(matches!(receipt, DeployAccount(_)));
+
+    let stdout_str = bdg.as_stdout();
+    assert!(stdout_str.contains("account deploy"));
+    assert!(stdout_str.contains("transaction_hash"));
+
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file)).unwrap();
+    let items: Value = serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
+    assert_eq!(items["alpha-sepolia"]["my_account"]["deployed"], true);
+}
+
+#[tokio::test]
+pub async fn test_happy_case_v3() {
+    let tempdir = create_account(false, &OZ_CLASS_HASH.into_hex_string(), "oz").await;
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--json",
+        "account",
+        "deploy",
+        "--name",
+        "my_account",
+        "--version",
+        "v3",
+        "--max-gas",
+        "1000",
+        "--max-gas-unit-price",
+        "100000000000",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+    let bdg = snapbox.assert();
+
+    let hash = get_transaction_hash(&bdg.get_output().stdout);
+    let receipt = get_transaction_receipt(hash).await;
+
+    assert!(matches!(receipt, DeployAccount(_)));
+
+    let stdout_str = bdg.as_stdout();
+    assert!(stdout_str.contains("account deploy"));
+    assert!(stdout_str.contains("transaction_hash"));
+
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file)).unwrap();
+    let items: Value = serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
+    assert_eq!(items["alpha-sepolia"]["my_account"]["deployed"], true);
+}
+
+#[tokio::test]
+pub async fn test_happy_case_strk_max_fee() {
+    let tempdir = create_account(false, &OZ_CLASS_HASH.into_hex_string(), "oz").await;
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--json",
+        "account",
+        "deploy",
+        "--name",
+        "my_account",
+        "--fee-token",
+        "strk",
+        "--max-fee",
+        "100000000000000",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+    let bdg = snapbox.assert();
+
+    let hash = get_transaction_hash(&bdg.get_output().stdout);
+    let receipt = get_transaction_receipt(hash).await;
+
+    assert!(matches!(receipt, DeployAccount(_)));
+
+    let stdout_str = bdg.as_stdout();
+    assert!(stdout_str.contains("account deploy"));
+    assert!(stdout_str.contains("transaction_hash"));
+
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file)).unwrap();
+    let items: Value = serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
     assert_eq!(items["alpha-sepolia"]["my_account"]["deployed"], true);
 }
 
@@ -76,6 +237,8 @@ pub async fn test_happy_case_add_profile() {
         "my_account",
         "--max-fee",
         "99999999999999999",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -110,6 +273,8 @@ fn test_account_deploy_error(accounts_content: &str, error: &str) {
         "my_account",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(temp_dir.path());
@@ -135,6 +300,8 @@ async fn test_too_low_max_fee() {
         "my_account",
         "--max-fee",
         "1",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -146,6 +313,64 @@ async fn test_too_low_max_fee() {
         command: account deploy
         error: Max fee is smaller than the minimal transaction cost
         "},
+    );
+}
+
+#[test_case("eth", "v3"; "eth-v3")]
+#[test_case("strk", "v1"; "strk-v3")]
+#[tokio::test]
+async fn test_invalid_version_and_token_combination(fee_token: &str, version: &str) {
+    let tempdir = create_account(false, &OZ_CLASS_HASH.into_hex_string(), "oz").await;
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--wait",
+        "account",
+        "deploy",
+        "--name",
+        "my_account",
+        "--fee-token",
+        fee_token,
+        "--version",
+        version,
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+
+    let output = snapbox.assert().failure();
+    assert_stderr_contains(
+        output,
+        format!("Error: {fee_token} fee token is not supported for {version} deployment."),
+    );
+}
+
+#[tokio::test]
+async fn test_no_version_and_token() {
+    let tempdir = create_account(false, &OZ_CLASS_HASH.into_hex_string(), "oz").await;
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--url",
+        URL,
+        "--accounts-file",
+        accounts_file,
+        "--wait",
+        "account",
+        "deploy",
+        "--name",
+        "my_account",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+
+    let output = snapbox.assert().failure();
+    assert_stderr_contains(
+        output,
+        "Error: Either --fee-token or --version must be provided",
     );
 }
 
@@ -165,6 +390,8 @@ pub async fn test_valid_class_hash() {
         "my_account",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -191,6 +418,8 @@ pub async fn test_valid_no_max_fee() {
         "deploy",
         "--name",
         "my_account",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -280,6 +509,8 @@ pub async fn test_happy_case_keystore(account_type: &str) {
         "deploy",
         "--max-fee",
         "99999999999999999",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -290,8 +521,7 @@ pub async fn test_happy_case_keystore(account_type: &str) {
     "});
 
     let contents = fs::read_to_string(tempdir.path().join(account_file)).unwrap();
-    let items: serde_json::Value =
-        serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
+    let items: Value = serde_json::from_str(&contents).expect("Failed to parse accounts file at ");
     assert_eq!(items["deployment"]["status"], "deployed");
     assert!(!items["deployment"]["address"].is_null());
     assert!(items["deployment"]["salt"].is_null());
@@ -326,6 +556,8 @@ pub async fn test_keystore_already_deployed() {
         "deploy",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -369,6 +601,8 @@ pub async fn test_keystore_key_mismatch() {
         "deploy",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -407,6 +641,8 @@ pub async fn test_deploy_keystore_inexistent_keystore_file() {
         "deploy",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -445,6 +681,8 @@ pub async fn test_deploy_keystore_inexistent_account_file() {
         "deploy",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -487,6 +725,8 @@ pub async fn test_deploy_keystore_no_status() {
         "deploy",
         "--max-fee",
         "10000000000000000",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());
@@ -542,6 +782,8 @@ pub async fn test_deploy_keystore_other_args() {
         "some-name",
         "--max-fee",
         "99999999999999999",
+        "--fee-token",
+        "eth",
     ];
 
     let snapbox = runner(&args).current_dir(tempdir.path());