+++ crates/sncast/tests/helpers/fixtures.rs
@@ -15,7 +15,7 @@ use sncast::state::state_file::{
 use sncast::{apply_optional, get_chain_id, get_keystore_password, AccountType};
 use sncast::{get_account, get_provider};
 use starknet::accounts::{
-    Account, AccountFactory, ArgentAccountFactory, Call, Execution, OpenZeppelinAccountFactory,
+    Account, AccountFactory, ArgentAccountFactory, Call, ExecutionV1, OpenZeppelinAccountFactory,
 };
 use starknet::core::types::TransactionReceipt;
 use starknet::core::types::{FieldElement, InvokeTransactionResult};
@@ -133,7 +133,7 @@ async fn deploy_oz_account(address: &str, class_hash: &str, salt: &str, private_
 async fn deploy_account_to_devnet<T: AccountFactory + Sync>(factory: T, address: &str, salt: &str) {
     mint_token(address, u64::MAX).await;
     factory
-        .deploy(salt.parse().expect("Failed to parse salt"))
+        .deploy_v1(salt.parse().expect("Failed to parse salt"))
         .send()
         .await
         .expect("Failed to deploy account");
@@ -202,8 +202,8 @@ pub async fn invoke_contract(
         calldata,
     };
 
-    let execution = account.execute(vec![call]);
-    let execution = apply_optional(execution, max_fee, Execution::max_fee);
+    let execution = account.execute_v1(vec![call]);
+    let execution = apply_optional(execution, max_fee, ExecutionV1::max_fee);
 
     execution.send().await.unwrap()
 }
@@ -212,19 +212,22 @@ pub async fn invoke_contract(
 // but serde_json cannot serialize numbers this big
 pub async fn mint_token(recipient: &str, amount: u64) {
     let client = reqwest::Client::new();
-    let json = json!(
-        {
-            "address": recipient,
-            "amount": amount
-        }
-    );
-    client
-        .post("http://127.0.0.1:5055/mint")
-        .header("Content-Type", "application/json")
-        .body(json.to_string())
-        .send()
-        .await
-        .expect("Error occurred while minting tokens");
+    for unit in ["FRI", "WEI"] {
+        let json = json!(
+            {
+                "address": recipient,
+                "amount": amount,
+                "unit": unit,
+            }
+        );
+        client
+            .post("http://127.0.0.1:5055/mint")
+            .header("Content-Type", "application/json")
+            .body(json.to_string())
+            .send()
+            .await
+            .expect("Error occurred while minting tokens");
+    }
 }
 
 #[must_use]